# docs/api/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "INOVAGAB — Águia Branca API",
    "description": "Plataforma de inovação corporativa: orientações estratégicas, ideias, projetos, relatórios e insights de IA.",
    "version": "v1"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/ideas/{id}": {
      "get": {
        "tags": [
          "Ideas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Ideas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdeaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IdeaRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IdeaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Ideas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/ideas/{id}/ice": {
      "put": {
        "tags": [
          "Ideas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IceRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IceRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ideas/{id}/reject": {
      "post": {
        "tags": [
          "Ideas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ideas/{id}/approve": {
      "post": {
        "tags": [
          "Ideas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApproveIdeaResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproveIdeaResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproveIdeaResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApproveIdeaResult": {
        "type": "object",
        "properties": {
          "ideaId": {
            "type": "string",
            "nullable": true
          },
          "projectId": {
            "type": "string",
            "nullable": true
          },
          "alreadyApproved": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "nullable": true
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          },
          "expiresIn": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfileResponse"
          }
        },
        "additionalProperties": false
      },
      "Division": {
        "enum": [
          "PASSAGEIROS",
          "COMERCIO",
          "LOGISTICA",
          "CORPORATIVO"
        ],
        "type": "string"
      },
...
      "IceRequest": {
        "type": "object",
        "properties": {
          "impact": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "confidence": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ease": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IceResponse": {
        "type": "object",
        "properties": {
          "impact": {
            "type": "integer",
            "format": "int32"
          },
          "confidence": {
            "type": "integer",
            "format": "int32"
          },
          "ease": {
            "type": "integer",
            "format": "int32"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "IdeaRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "division": {
            "$ref": "#/components/schemas/Division"
          },
          "guidelineId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IdeaResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "division": {
            "$ref": "#/components/schemas/Division"
          },
          "guidelineId": {
            "type": "string",
            "nullable": true
          },
          "guidelineTitle": {
            "type": "string",
            "nullable": true
          },
          "authorId": {
            "type": "string",
            "nullable": true
          },
          "authorName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/IdeaStatus"
          },
          "ice": {
            "$ref": "#/components/schemas/IceResponse"
          },
          "reviewerId": {
            "type": "string",
            "nullable": true
          },
          "reviewComment": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "linkedProject": {
            "$ref": "#/components/schemas/LinkedProjectResponse"
          },
          "pointsAwarded": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "IdeaStatus": {
        "enum": [
          "SUBMETIDA",
          "EM_ANALISE",
          "APROVADA",
          "REJEITADA",
          "IMPLEMENTADA"
        ],
        "type": "string"
      },
...
      "LinkedProjectResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "stage": {
            "$ref": "#/components/schemas/ProjectStage"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LogoutRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProjectStage": {
        "enum": [
          "PLANEJAMENTO",
          "EM_EXECUCAO",
          "CONCLUIDO",
          "CANCELADO"
        ],
        "type": "string"
      },
...
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RejectRequest": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Role": {
        "enum": [
          "OPERADOR",
          "GESTOR",
          "LIDER"
        ],
        "type": "string"
      },
...
      "UserProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "division": {
            "$ref": "#/components/schemas/Division"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          },
          "badges": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Informe o accessToken JWT obtido em /api/v1/auth/login.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}